Added timestamps

+++ Models/Chat.cs
@@ -4,7 +4,7 @@ namespace WordsCountBot.Models
 {
     public class Chat : Model
     {
-        public int TelegramID { get; set; }
+        public long TelegramID { get; set; }
         public string Name { get; set; }
         public IList<WordUsedTimes> Usages { get; set; }
     }
